updates the file names, add comments in the code
