+++ package.json
@@ -12,11 +12,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "body-parser": "^1.20.1",
-    "cors": "^2.8.5",
     "csvtojson": "^2.0.10",
+    "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "mongoose": "^6.8.3",
-    "multer": "^1.4.5-lts.1"
+    "mongoose": "^6.8.3"
   }
 }
