+++ src/app.js
@@ -1,23 +1,26 @@
 const express = require("express");
-require("./db/conn");
-const PArouting = require("./routers/Pain_Area_router");
-const PDrouting = require("./routers/Pain_Definition_router");
-const PBrouting = require("./routers/Pain_Behavior_router");
-const QuesRouting = require("./routers/question_router");
-const Diagrouting = require("./routers/Diagnostics_router");
-const PBQrouting = require("./routers/PainBehaviorQuestion_router");
-const PPDrouting = require("./routers/PainPossibleDiagnotics_router");
-const Probrouting = require("./routers/Probability_router");
-const AssignResultrouting = require("./routers/AssginResult_router");
-const DiagResultrouting = require("./routers/DaignonsisResult_router");
+const mongoose = require("mongoose");
+const config = require("./config/config");
+require("./config/db");
+
+
+const PArouting = require("./routers/painArea");
+const PDrouting = require("./routers/painDefinition");
+const PBrouting = require("./routers/painBehavior");
+const QuesRouting = require("./routers/question");
+const Diagrouting = require("./routers/diagnostics");
+const PBQrouting = require("./routers/painBehaviorQuestion");
+const PPDrouting = require("./routers/painPossibleDiagnotics");
+const Probrouting = require("./routers/probability");
+const AssignResultrouting = require("./routers/assginResult");
+const DiagResultrouting = require("./routers/daignonsisResult");
 
 const app = express();
-const PORT = process.env.PORT || 80;
 
 app.use(express.json());
 app.use([PArouting, PDrouting, PBrouting, QuesRouting, Diagrouting, PBQrouting, PPDrouting, Probrouting, AssignResultrouting, DiagResultrouting]);
 
 
-app.listen(PORT, () => {
-    console.log(`Server is running at port no ${PORT}`);
+app.listen(config.PORT, () => {
+    console.log(`Server is running at port no ${config.PORT}`);
 })
