--- src/db/conn.js
@@ -1,9 +1,0 @@
-const mongoose = require("mongoose");
-
-mongoose.set('strictQuery', true);
-mongoose.connect("mongodb://localhost:27017/sunndio")
-.then(() => {
-    console.log("Database Connection is done...");
-}).catch((err) => {
-    console.log(err);
-});
