+++ src/models/painPossibleDiagnostics.js
@@ -1,6 +1,7 @@
 const mongoose = require("mongoose");
 
-const PPDSchema = new mongoose.Schema({
+// Define the mongoose Schema for painPossibleDiagnostic, it contained the painBehaviorId, diagnosticsId
+const painPossibleDiagnosticSchema = new mongoose.Schema({
     Diagnostics_id : {
         type : mongoose.Schema.Types.ObjectId,
         ref : "diagnostic"
@@ -14,4 +15,5 @@ const PPDSchema = new mongoose.Schema({
     }
 });
 
-module.exports = mongoose.model("PainPossibleDiagnotic", PPDSchema);;
+// export the painPossibleDiagnostic model
+module.exports = mongoose.model("PainPossibleDiagnotic", painPossibleDiagnosticSchema);;
